orders_service/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
asyncpg>=0.24.0
tenacity>=8.0.0
pydantic>=1.8.0
httpx==0.23.0
aiokafka>=0.12.0
aioredis>=2.0.1
python-jose[cryptography]>=3.3.0
alembic>=1.12
psycopg2-binary>=2.9
prometheus-fastapi-instrumentator
python-json-logger>=3.3.0